refactor(utils): add getDiffInCalendarQuarters as ESM
