+++ src/utils/index.ts
@@ -5,4 +5,5 @@ export { default as getDiffInCalendarHours } from './getDiffInCalendarHours.js';
 export { default as getDiffInCalendarDays } from './getDiffInCalendarDays.js';
 export { default as getDiffInCalendarWeeks } from './getDiffInCalendarWeeks.js';
 export { default as getDiffInCalendarMonths } from './getDiffInCalendarMonths.js';
+export { default as getDiffInCalendarQuarters } from './getDiffInCalendarQuarters.js';
 export { default as getDiffInCalendarYears } from './getDiffInCalendarYears.js';
